api/requirements.txt
# --- API & Web Server ---
fastapi
uvicorn
python-dotenv
streamlit
python-multipart

# --- AI & RAG Orchestration ---
neo4j
langgraph
langchain
langchain-openai
langchain-community
langchain-ollama

# CDM
commondatamodel-objectmodel

# --- Parsing etc. ---
pypdf
pydantic-settings
pyyaml
pytest
pyprojroot
